^ [#15886] Merged revisions 2360-2363 from /branches/1.6-xillibit with minor changes
git-svn-id: file:///Users/oratzes/svn/kunena@2364 daf2aa89-cb5b-0410-b8ee-fe93ebc9b27d

+++ trunk/1.6/components/com_kunena/funcs/profile.php
@@ -404,9 +404,11 @@ class CKunenaProfile {
 			$path = KUNENA_PATH_AVATAR_UPLOADED .DS. $uploadpath;
 
 			// Delete old uploaded avatars:
-			$deletelist = JFolder::files($path, 'user'.$this->profile->userid, false, true);
-			foreach ($deletelist as $delete) {
-				JFile::delete($delete);
+			if ( JFolder::exists( $path ) ) {
+				$deletelist = JFolder::files($path, 'user'.$this->profile->userid, false, true);
+				foreach ($deletelist as $delete) {
+					JFile::delete($delete);
+				}
 			}
 			$upload->setImageResize(intval($this->config->avatarsize)*1024, 200, 200, $this->config->avatarquality);
